refactor: 가이드 페이지 디렉토리 이동, api, 배포 메뉴는 만드려 했는데 보류
